refactor: move ai provider catalog logic into core
