+++ crates/commit-cat-core/src/models/mod.rs
@@ -1,4 +1,5 @@
-pub mod cat;
 pub mod activity;
-pub mod settings;
+pub mod ai_provider_catalog;
+pub mod cat;
 pub mod growth;
+pub mod settings;
